tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.readonly.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.readonly</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.readonly</h1>
<p class="desc">CanvasRenderingContext2D.canvas is readonly</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("CanvasRenderingContext2D.canvas is readonly");
_addTest(function(canvas, ctx) {

  var c = document.createElement('canvas');
  var d = ctx.canvas;
  _assertDifferent(c, d, "c", "d");
  ctx.canvas = c;
  _assertSame(ctx.canvas, d, "ctx.canvas", "d");

});
</script>
