LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/canvas-host/2d.canvas.host.readonly-expected.txt

2d.canvas.host.readonly
CanvasRenderingContext2D.canvas is readonly
Actual output:

PASS CanvasRenderingContext2D.canvas is readonly
